Fix nasty bug in MovieGenreList.findListByMovieId
In javascript, one shouldn't use array.forEach with async functions, but no errors occur
if you do, and race conditions start happening. In this case, the resulting genres list
was empty for some movies, but correct for other movies. It took me more than one hour to
debug the issue.